changin names and structure and adding profile
