+++ src/router/index.js
@@ -2,7 +2,8 @@ import Vue from 'vue'
 import Router from 'vue-router'
 import Home from '@/pages/Home/Home'
 import ProjectsContainer from '@/pages/Projects/ProjectsContainer'
-import Profiles from '@/pages/Profiles/Profiles'
+import Community from '@/pages/Community/Community'
+import Profile from '@/pages/Profile/Profile'
 import Events from '@/pages/Events/Events'
 import Contact from '@/pages/Contact/Contact'
 import History from '@/pages/History/History'
@@ -24,10 +25,15 @@ export default new Router({
       component: ProjectsContainer
     },
     {
-      path: '/profiles',
-      name: 'Profiles',
+      path: '/community',
+      name: 'Community',
       props: true,
-      component: Profiles
+      component: Community
+    },
+    {
+      path: '/profile',
+      name: 'Profile',
+      component: Profile
     },
     {
       path: '/events',
